Seperated data generators from web calls
